On Feb 2026

+++ IPSR/Numpy/5_array_properties.py
@@ -1,6 +1,8 @@
 import numpy as np
-arr = np.array([[1, 2, 3], [4, 5, 6]],dtype=np.int32)
+arr = np.array([[1, 2, 3], [4, 5, 6]],dtype=np.int64)
 print("Shape:", arr.shape)
 print("Size:", arr.size)
 print("Dimension:", arr.ndim)
 print("Datatype:", arr.dtype)
+
+
